test: harden reusable vitest infrastructure
