+++ packages/test/setup-server.js
@@ -1,0 +1,1 @@
+module.exports = require('./lib/setup-server');
