+++ packages/test/src/__tests__/server-entry-export.test.ts
@@ -1,9 +1,9 @@
-import { setupServerTestEnvironment } from '@tachybase/test/server';
+import { setupServerTestEnvironment } from '@tachybase/test/setup-server';
 
 import { describe, expect, it } from 'vitest';
 
-describe('@tachybase/test/server export', () => {
-  it('exports setupServerTestEnvironment from the server entry', () => {
+describe('@tachybase/test/setup-server export', () => {
+  it('exports setupServerTestEnvironment from the setup server entry', () => {
     expect(setupServerTestEnvironment).toEqual(expect.any(Function));
   });
 });
