+++ packages/test/src/setup-server.ts
@@ -1,0 +1,1 @@
+export * from './server/setupTestEnvironment';
