+++ packages/test/vitest.d.ts
@@ -1,0 +1,2 @@
+export * from './lib/vitest';
+export { default } from './lib/vitest';
